Updating DataModels for HTTP cleanup-TEW
