+++ dotnet/src/SemanticKernel.Skills/Skills.Memory.Qdrant/Internal/HttpSchema/DeleteVectorsRequest.cs
@@ -3,7 +3,8 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Text.Json.Serialization;
-using Microsoft.SemanticKernel.Skills.Memory.VectorDB.Internal.Diagnostics;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.Diagnostics;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
 
 namespace Micrsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
