+++ dotnet/src/SemanticKernel.Skills/Skills.Memory.Qdrant/Internal/HttpSchema/FetchCollectionRequest.cs
@@ -1,6 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System.Net.Http;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.Diagnostics;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
 using Microsoft.SemanticKernel.Skills.Memory.VectorDB.Internal.Diagnostics;
 
 namespace Micrsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
