+++ dotnet/src/SemanticKernel.Skills/Skills.Memory.Qdrant/Internal/HttpSchema/FetchCollectionsNameRequest.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using System.Net.Http;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
 using Microsoft.SemanticKernel.Skills.Memory.Qdrant.SDKClient.Internal.Diagnostics;
 
 namespace Microsoft.SemanticKernel.Skills.Memory.Qdrant.SDKClient.Internal.HttpSchema;
@@ -17,11 +18,6 @@ internal class FetchCollectionNameRequest
         return HttpRequest.CreateGetRequest($"collections");
     }
 
-    #region private ================================================================================
-    private FetchAllCollectionNameRequest()
-    {
-    }
-
-    #endregion  
+    
 
 }
