+++ dotnet/src/SemanticKernel.Skills/Skills.Memory.Qdrant/Internal/HttpSchema/FetchVectorsRequest.cs
@@ -3,7 +3,9 @@
 using System.Collections.Generic;
 using System.Net.Http;
 using System.Text.Json.Serialization;
-using Microsoft.SemanticKernel.Skills.Memory.VectorDB.Internal.Diagnostics;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.DataModels;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.Diagnostics;
+using Microsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
 
 namespace Micrsoft.SemanticKernel.Skills.Memory.Qdrant.HttpSchema;
 
@@ -52,74 +54,6 @@ internal class FetchVectorsRequest : IValidatable
         this.Filters = new Filter();
         this.WithPayload = true;
         this.WithVector = true;
-    }
-
-    private class Filter : IValidatable
-    {
-        internal class Match : IValidatable
-        {
-            [JsonPropertyName("value")]
-            internal object Value { get; set; }
-
-            public Match()
-            {
-                this.Value = string.Empty;
-            }
-
-            public void Validate()
-            {
-            }
-        }
-
-        internal class Must : IValidatable
-        {
-            [JsonPropertyName("key")]
-            internal string Key { get; set; }
-
-            [JsonPropertyName("match")]
-            internal Match Match { get; set; }
-
-            public Must()
-            {
-                this.Key = string.Empty;
-                this.Match = new();
-            }
-
-            public Must(string key, object value) : this()
-            {
-                this.Key = key;
-                this.Match.Value = value;
-            }
-
-            public void Validate()
-            {
-                Verify.NotNullOrEmpty(this.Key, "The filter key is empty");
-                Verify.NotNull(this.Match, "The filter match is NULL");
-            }
-        }
-
-        [JsonPropertyName("must")]
-        internal List<Must> Conditions { get; set; }
-
-        internal Filter()
-        {
-            this.Conditions = new();
-        }
-
-        internal Filter ValueMustMatch(string key, object value)
-        {
-            this.Conditions.Add(new Must(key, value));
-            return this;
-        }
-
-        public void Validate()
-        {
-            Verify.NotNull(this.Conditions, "Filter conditions are NULL");
-            foreach (var x in this.Conditions)
-            {
-                x.Validate();
-            }
-        }
     }
 
     #endregion
